+++ components/Pages/GrantMilestonesAndUpdates/screens/index.ts
@@ -2,4 +2,3 @@ export * from "./EmptyGrantsSection";
 export * from "./MilestonesAndUpdates";
 export * from "./NewGrant";
 export * from "./NewGrantUpdate";
-export * from "./NewMilestone";
